requirements.txt
pillow
psutil
pyinstaller 